Refactored the method for asset fetching into separate extension
