+++ ownCloud/Tasks/InstantMediaUploadTaskExtension.swift
@@ -65,7 +65,7 @@ class InstantMediaUploadTaskExtension : ScheduledTaskAction {
 
 		// Add photo assets
 		if let uploadPhotosAfter = userDefaults.instantUploadPhotosAfter {
-			let fetchResult = self.fetchAssetsFromCameraRoll(with: [.image], createdAfter: uploadPhotosAfter)
+			let fetchResult = PHAsset.fetchAssetsFromCameraRoll(with: [.image], createdAfter: uploadPhotosAfter)
 			if fetchResult != nil {
 				fetchResult!.enumerateObjects({ (asset, _, _) in
 					photoAssets.append(asset)
@@ -91,7 +91,7 @@ class InstantMediaUploadTaskExtension : ScheduledTaskAction {
 
 		// Add video assets
 		if let uploadVideosAfter = userDefaults.instantUploadVideosAfter {
-			let fetchResult = self.fetchAssetsFromCameraRoll(with: [.video], createdAfter: uploadVideosAfter)
+			let fetchResult = PHAsset.fetchAssetsFromCameraRoll(with: [.video], createdAfter: uploadVideosAfter)
 			if fetchResult != nil {
 				fetchResult!.enumerateObjects({ (asset, _, _) in
 					videoAssets.append(asset)
@@ -106,50 +106,6 @@ class InstantMediaUploadTaskExtension : ScheduledTaskAction {
 			userDefaults.instantUploadVideosAfter = videoAssets.last?.creationDate
             Log.debug(tagged: ["INSTANT_MEDIA_UPLOAD"], "Last added video asset modification date: \(String(describing: userDefaults.instantUploadPhotosAfter))")
 		}
-	}
-
-	private func fetchAssetsFromCameraRoll(with mediaTypes:[PHAssetMediaType], createdAfter:Date? = nil) -> PHFetchResult<PHAsset>? {
-
-		guard PHPhotoLibrary.authorizationStatus() == .authorized else { return nil }
-
-		let collectionResult = PHAssetCollection.fetchAssetCollections(with: .smartAlbum,
-																	   subtype: .smartAlbumUserLibrary,
-																	   options: nil)
-
-		if let cameraRoll = collectionResult.firstObject {
-			let imageTypePredicate = NSPredicate(format: "mediaType = %d", PHAssetMediaType.image.rawValue)
-			let videoTypePredicate = NSPredicate(format: "mediaType = %d", PHAssetMediaType.video.rawValue)
-
-			var typePredicatesArray = [NSPredicate]()
-
-			if mediaTypes.contains(.image) {
-				typePredicatesArray.append(imageTypePredicate)
-			}
-
-			if mediaTypes.contains(.video) {
-				typePredicatesArray.append(videoTypePredicate)
-			}
-
-			let mediaTypesPredicate = NSCompoundPredicate(orPredicateWithSubpredicates: typePredicatesArray)
-
-			let fetchOptions = PHFetchOptions()
-
-			if let date = createdAfter {
-				let creationDatePredicate = NSPredicate(format: "creationDate > %@", date as NSDate)
-				fetchOptions.predicate = NSCompoundPredicate(andPredicateWithSubpredicates: [mediaTypesPredicate, creationDatePredicate])
-			} else {
-				fetchOptions.predicate = mediaTypesPredicate
-			}
-
-			let sort = NSSortDescriptor(key: "creationDate", ascending: true)
-			fetchOptions.sortDescriptors = [sort]
-
-            Log.debug(tagged: ["INSTANT_MEDIA_UPLOAD"], "Fetching assets with options \(fetchOptions.debugDescription)")
-
-			return PHAsset.fetchAssets(in: cameraRoll, options: fetchOptions)
-		}
-
-		return nil
 	}
 
 	private func showFeatureDisabledAlert() {
